Sinc manual cambios Francisco Corrección tipos de dato decimal y fecha

+++ src/main/java/epntech/cbdmq/pe/dominio/admin/PruebaFisica.java
@@ -1,16 +1,22 @@
 package epntech.cbdmq.pe.dominio.admin;
 
-import java.math.BigDecimal;
+import java.util.ArrayList;
+import java.util.List;
 
 import org.hibernate.annotations.ResultCheckStyle;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.Where;
 
+import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.JoinTable;
+import jakarta.persistence.ManyToMany;
 import jakarta.persistence.Table;
 import lombok.Data;
 
@@ -36,11 +42,11 @@ public class PruebaFisica {
 	@Column(name = "sexo")
 	private String sexo;
 	@Column(name = "calificacion")
-	private BigDecimal calificacion;
+	private Integer calificacion;
 	@Column(name = "peso_prueba")
-	private BigDecimal pesoprueba;
+	private Integer pesoprueba;
 	@Column(name = "valor")
-	private BigDecimal valor;
+	private Integer valor;
 	@Column(name = "no_min_flexion")
 	private Integer nominflexion;
 	@Column(name = "estado")
